stridx's prototype has size_t as return type, while the function had unsigned as return type. This has been corrected by setting them both to size_t.
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@1122 7b491191-dbf0-0310-aff6-d879d4d69008

+++ imaputil.c
@@ -1913,9 +1913,9 @@ char *date_imap2sql(const char *imapdate)
 /*
  *
  */
-unsigned stridx(const char *s, char ch)
+size_t stridx(const char *s, char ch)
 {
-	unsigned i;
+	size_t i;
 
 	for (i = 0; s[i] && s[i] != ch; i++);
